backend/requirements.txt
# existing...
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
PyYAML==6.0.2
python-dateutil==2.9.0.post0
scikit-learn==1.5.1
joblib==1.4.2
reportlab==4.2.2

# new for v2
sentence-transformers==3.0.1
hdbscan==0.8.38.post1
umap-learn==0.5.6     # optional, speeds up cluster separation
faiss-cpu==1.12.0
numpy==1.26.4
pandas==2.2.2